+++ src/roq/samples/example-9/config.hpp
@@ -1,0 +1,32 @@
+/* Copyright (c) 2017-2025, Hans Erik Thrane */
+
+#pragma once
+
+#include "roq/api.hpp"
+#include "roq/client.hpp"
+
+#include "roq/samples/example-9/settings.hpp"
+
+namespace roq {
+namespace samples {
+namespace example_9 {
+
+// client::Config is an interface allowing a strategy factory
+// to query for relevant configuration
+// the design uses dispatch to allow the factory to use interfaces
+// without imposing any kind of requirement on containers and storage
+
+struct Config final : public client::Config {
+  explicit Config(Settings const &);
+
+  Config(Config &&) = default;
+  Config(Config const &) = delete;
+
+ protected:
+  Settings const &settings_;
+  void dispatch(Handler &) const override;
+};
+
+}  // namespace example_9
+}  // namespace samples
+}  // namespace roq