+++ src/roq/samples/example-9/main.cpp
@@ -1,0 +1,28 @@
+/* Copyright (c) 2017-2025, Hans Erik Thrane */
+
+#include "roq/api.hpp"
+
+#include "roq/flags/args.hpp"
+#include "roq/logging/flags/settings.hpp"
+
+#include "roq/samples/example-9/application.hpp"
+
+using namespace std::literals;
+
+// === CONSTANTS ===
+
+namespace {
+auto const INFO = roq::Service::Info{
+    .description = "Example 9 (Roq Samples)"sv,
+    .package_name = ROQ_PACKAGE_NAME,
+    .build_version = ROQ_VERSION,
+};
+}  // namespace
+
+// === IMPLEMENTATION ===
+
+int main(int argc, char **argv) {
+  roq::flags::Args args{argc, argv, INFO.description, INFO.build_version};
+  roq::logging::flags::Settings settings{args};
+  return roq::samples::example_9::Application{args, settings, INFO}.run();
+}