+++ src/roq/samples/example-9/settings.cpp
@@ -1,0 +1,16 @@
+/* Copyright (c) 2017-2025, Hans Erik Thrane */
+
+#include "roq/samples/example-9/settings.hpp"
+
+#include "roq/client/flags/settings.hpp"
+
+namespace roq {
+namespace samples {
+namespace example_9 {
+
+Settings::Settings(args::Parser const &args) : client::flags::Settings{args}, flags::Flags{flags::Flags::create()} {
+}
+
+}  // namespace example_9
+}  // namespace samples
+}  // namespace roq