+++ src/roq/samples/example-9/test.sh
@@ -1,0 +1,26 @@
+#!/usr/bin/env bash
+
+NAME="trader"
+
+# debug?
+
+if [ "$1" == "debug" ]; then
+  KERNEL="$(uname -a)"
+  case "$KERNEL" in
+    Linux*)
+      PREFIX="gdb --args"
+      ;;
+    Darwin*)
+      PREFIX="lldb --"
+      ;;
+  esac
+  shift 1
+else
+  PREFIX=
+fi
+
+# launch
+
+$PREFIX "./roq-cpp-samples-example-9" \
+  --name "$NAME" \
+  $@